requirement.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
lxml==6.0.0
pillow==11.3.0
python-pptx==1.0.2
requests==2.32.4
typing_extensions==4.14.1
urllib3==2.5.0
xlsxwriter==3.2.5
